Clean up db access code
